+++ application/models/Transactions.php
@@ -1,11 +1,9 @@
 <?php
 
-
 class Transactions extends MY_Model{
     public function __construct()
     {
-        parent::__construct('Transactions', 'transactionID');
+        parent::__construct('transactions', 'transactionID');
     }
-
 
 }